# user-server/docs/swagger.yaml
basePath: /
definitions:
...
  hivemtk-user_internal_service.LowRecallQuery:
    properties:
      created_at:
        type: string
      hit_count:
        type: integer
      id:
        type: integer
      latency_ms:
        type: integer
      precision:
        type: number
      query:
        type: string
      recall:
        type: number
      relevant_count:
        type: integer
      retrieved_count:
        type: integer
      session_id:
        type: string
    type: object
...
  hivemtk-user_internal_service.RagRecallMetricsSummary:
    properties:
      avg_latency_ms:
        type: number
      avg_precision:
        type: number
      avg_recall:
        type: number
      avg_similarity:
        type: number
      low_recall_count:
        type: integer
      p95_latency_ms:
        type: integer
      top_1_hit_rate:
        type: number
      top_k_hit_rate:
        type: number
      total_queries:
        type: integer
      window_end:
        type: string
      window_start:
        type: string
      zero_hit_count:
        type: integer
    type: object
...
  internal_controller.CollectRequest:
    properties:
      window_seconds:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    HiveMtk 用户端（私域独立部署）RESTful API。认证采用 JWT：请求头 Authorization: Bearer <token>。
    仅当 ENABLE_SWAGGER=true 时，通过 http://localhost:8204/swagger/index.html 本地访问。
  title: HiveMtk 用户端 API
  version: 3.41.0
paths:
...
  /api/rag/recall/collect:
    post:
      consumes:
      - application/json
      description: 按指定时间窗口（默认 1 小时）聚合 rag_query_logs 并写入监控快照表
      parameters:
      - description: 采集参数
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_controller.CollectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hivemtk-user_internal_service.RagRecallMetricsSummary'
      summary: 手动触发一次召回率指标采集
      tags:
      - RAG Recall
  /api/rag/recall/low-recall:
    get:
      description: 用于调优分析，按创建时间倒序返回低召回查询
      parameters:
      - description: 召回率阈值（默认 0.3）
        in: query
        name: threshold
        type: number
      - description: 返回条数（默认 50，最大 200）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hivemtk-user_internal_service.LowRecallQuery'
            type: array
      summary: 查询召回率低于阈值的样本
      tags:
      - RAG Recall
...
  /api/rag/recall/snapshots:
    get:
      description: 按时间窗口倒序返回历史快照（默认 50，上限 1000）
      parameters:
      - description: 返回条数（默认 50，最大 1000）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: 列出最近 N 条监控快照
      tags:
      - RAG Recall
...
  /api/system/init-complete:
    post:
      description: 标记初始化向导完成（写 install.lock.initialized=true，使 state 从 HAS_ADMIN 推进到
        INITIALIZED）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
            type: object
      summary: 完成系统初始化向导
      tags:
      - 系统初始化
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"